10/EPMDX Review
The resource path is a slightly different beast than MQTT topic
name. Resource path is defined in 1/KP and is an abstraction for
hiding MQTT/CoAP differences. When we talk about request/response, the
resource path is the same for both requests and responses, MQTT topic
names are different, and CoAP `URI-Path`s are equal. `/status`
appending is MQTT-specific and 1/KP extensions should be independ from
underlying 1/KP protocol, so I removed that part.

Recent additions standardize the meaning of request `id` field, so I
adapt this protocol to the new conventions.

Also, fix some grammar, remove TOC (as it will be auto-generated), and
add 10/EPMDX to the RFC index.