1/KP Move Request ID to MQTT topic name
KP can not make assumptions about message payload format, so it is
impossible to _define_ generic rules for passing IDs and response
fields within the message payload.

Move Request ID to the topic name. Add a response type to the topic
name and define payload format for the error response type.