10/EPMDX Review
"Metadata keys" and "Metadata values" sections are not initial
requirements to the protocol and are a part of the design. Move them
to the "Design" section.

Add the reasoning behind the metadata key format.

Specify JSON encoding (UTF-8) and rephrase the sentence to identify
which file it refers to.

Add myself to contributors. Change Volodymyr's address to the new
domain.